refactor: rename `generate` to `generation`
